feat!: turn IceFlow into a compiled shared library
