+++ include/iceflow/consumer.hpp
@@ -33,69 +33,27 @@ class IceflowConsumer {
 public:
   IceflowConsumer(std::shared_ptr<IceFlow> iceflow, const std::string &subTopic,
                   uint32_t numberOfPartitions, uint32_t consumerPartitionIndex,
-                  uint32_t totalNumberOfConsumers)
-      : m_iceflow(iceflow), m_subTopic(subTopic),
-        m_numberOfPartitions(numberOfPartitions),
-        m_consumerPartitionIndex(consumerPartitionIndex),
-        m_totalNumberOfConsumers(totalNumberOfConsumers) {
+                  uint32_t totalNumberOfConsumers);
 
-    setTopicPartitions(m_numberOfPartitions, m_consumerPartitionIndex,
-                       m_totalNumberOfConsumers);
-  }
+  ~IceflowConsumer();
 
-  ~IceflowConsumer() { unsubscribeFromAllPartitions(); }
-
-  std::vector<uint8_t> receiveData() { return m_inputQueue.waitAndPopValue(); }
+  std::vector<uint8_t> receiveData();
 
   /**
    * Indicates whether the queue of this IceflowConsumer contains data.
    */
-  bool hasData() { return !m_inputQueue.empty(); }
+  bool hasData();
 
-  void setNumberOfPartitions(uint32_t numberOfPartitions) {
-    setTopicPartitions(numberOfPartitions, m_consumerPartitionIndex,
-                       m_totalNumberOfConsumers);
-    m_numberOfPartitions = numberOfPartitions;
-  }
+  void setNumberOfPartitions(uint32_t numberOfPartitions);
 
-  void setConsumerPartitionIndex(uint32_t consumerPartitionIndex) {
-    setTopicPartitions(m_numberOfPartitions, consumerPartitionIndex,
-                       m_totalNumberOfConsumers);
-    m_consumerPartitionIndex = consumerPartitionIndex;
-  }
+  void setConsumerPartitionIndex(uint32_t consumerPartitionIndex);
 
-  void setTotalNumberOfConsumers(uint32_t totalNumberOfConsumers) {
-    setTopicPartitions(m_numberOfPartitions, m_consumerPartitionIndex,
-                       totalNumberOfConsumers);
-    m_totalNumberOfConsumers = totalNumberOfConsumers;
-  }
+  void setTotalNumberOfConsumers(uint32_t totalNumberOfConsumers);
 
 private:
   void validatePartitionConfiguration(uint32_t numberOfPartitions,
                                       uint32_t consumerPartitionIndex,
-                                      uint32_t totalNumberOfConsumers) {
-    if (numberOfPartitions == 0) {
-      throw std::invalid_argument(
-          "At least one topic partition has to be defined!");
-    }
-
-    if (totalNumberOfConsumers <= consumerPartitionIndex) {
-      throw std::invalid_argument(
-          "The total number of consumers has to be larger "
-          "than the consumerPartitionIndex.");
-    }
-
-    if (numberOfPartitions < totalNumberOfConsumers) {
-      throw std::invalid_argument(
-          "The numberOfPartitions has to be at least as large as the "
-          "totalNumberOfConsumers.");
-    }
-
-    if (numberOfPartitions <= consumerPartitionIndex) {
-      throw std::invalid_argument("The numberOfPartitions has to be at least "
-                                  "as large as the consumerPartitionIndex.");
-    }
-  }
+                                      uint32_t totalNumberOfConsumers);
 
   /**
    * Updates the topic partitions this IceflowConsumer is subscribed to.
@@ -105,65 +63,11 @@ private:
    */
   void setTopicPartitions(uint32_t numberOfPartitions,
                           uint32_t consumerPartitionIndex,
-                          uint32_t totalNumberOfConsumers) {
-    validatePartitionConfiguration(numberOfPartitions, consumerPartitionIndex,
-                                   totalNumberOfConsumers);
+                          uint32_t totalNumberOfConsumers);
 
-    // TODO: Document that consumer indexes must start at 0
-    std::unordered_set<uint32_t> topicPartitions;
-    for (uint32_t i = consumerPartitionIndex; i < numberOfPartitions;
-         i += totalNumberOfConsumers) {
-      topicPartitions.emplace_hint(topicPartitions.end(), i);
-    }
+  uint32_t subscribeToTopicPartition(uint64_t topicPartition);
 
-    if (auto validIceflow = m_iceflow.lock()) {
-      std::vector<decltype(m_subscriptionHandles)::key_type> handlesToErase;
-      for (auto subscriptionHandle : m_subscriptionHandles) {
-        auto topicPartition = subscriptionHandle.first;
-
-        if (!topicPartitions.contains(topicPartition)) {
-          validIceflow->unsubscribe(subscriptionHandle.second);
-          handlesToErase.push_back(subscriptionHandle.first);
-        }
-      }
-
-      for (auto handleToErase : handlesToErase) {
-        m_subscriptionHandles.erase(handleToErase);
-      }
-
-      for (auto topicPartition : topicPartitions) {
-        if (!m_subscriptionHandles.contains(topicPartition)) {
-          auto subscriptionHandle = subscribeToTopicPartition(topicPartition);
-          m_subscriptionHandles.emplace(topicPartition, subscriptionHandle);
-        }
-      }
-    } else {
-      throw std::runtime_error("Iceflow instance has already expired.");
-    }
-  }
-
-  uint32_t subscribeToTopicPartition(uint64_t topicPartition) {
-    if (auto validIceflow = m_iceflow.lock()) {
-
-      std::function<void(std::vector<uint8_t>)> pushDataCallback =
-          std::bind(&RingBuffer<std::vector<uint8_t>>::push, &m_inputQueue,
-                    std::placeholders::_1);
-
-      return validIceflow->subscribeToTopicPartition(m_subTopic, topicPartition,
-                                                     pushDataCallback);
-    } else {
-      throw std::runtime_error("Iceflow instance has already expired.");
-    }
-  }
-
-  void unsubscribeFromAllPartitions() {
-    if (auto validIceflow = m_iceflow.lock()) {
-      for (auto subscriptionHandle : m_subscriptionHandles) {
-        validIceflow->unsubscribe(subscriptionHandle.second);
-      }
-    }
-    m_subscriptionHandles.clear();
-  }
+  void unsubscribeFromAllPartitions();
 
 private:
   const std::weak_ptr<IceFlow> m_iceflow;
