+++ include/iceflow/iceflow.hpp
@@ -64,86 +64,12 @@ public:
    * custom `face`.
    */
   IceFlow(const std::string &syncPrefix, const std::string &nodePrefix,
-          ndn::Face &face)
-      : m_syncPrefix(syncPrefix), m_nodePrefix(nodePrefix), m_face(face) {
-    ndn::svs::SecurityOptions secOpts(m_keyChain);
-
-    ndn::svs::SVSPubSubOptions opts;
-
-    m_svsPubSub = std::make_shared<ndn::svs::SVSPubSub>(
-        ndn::Name(m_syncPrefix), ndn::Name(m_nodePrefix), m_face,
-        std::bind(&IceFlow::onMissingData, this, _1), opts, secOpts);
-  }
+          ndn::Face &face);
 
 public:
-  void run() {
-    if (m_running) {
-      throw std::runtime_error("Iceflow instance is already running!");
-    }
+  void run();
 
-    std::thread svsThread([this] {
-      while (true) {
-        try {
-          m_face.processEvents(ndn::time::milliseconds(5000), true);
-        } catch (std::exception &e) {
-          NDN_LOG_ERROR("Error in event handling loop: " << e.what());
-        }
-      }
-    });
-
-    m_running = true;
-    while (m_running) {
-      NDN_LOG_INFO("Checking if producers are registered...");
-      std::unique_lock lock(m_producerRegistrationMutex);
-      m_producerRegistrationConditionVariable.wait(
-          lock, [this] { return m_producersAvailable; });
-      NDN_LOG_INFO("At least one producer is registered, continuing "
-                   "publishing thread.");
-
-      auto closestNextPublishTimePoint =
-          std::chrono::time_point<std::chrono::steady_clock>::max();
-
-      for (auto producerRegistrationTuple : m_producerRegistrations) {
-        auto producerRegistration = std::get<1>(producerRegistrationTuple);
-
-        auto nextPublishTimePoint =
-            producerRegistration.getNextPublishTimePoint();
-        auto timeUntilNextPublish =
-            nextPublishTimePoint - std::chrono::steady_clock::now();
-
-        if (nextPublishTimePoint < closestNextPublishTimePoint) {
-          closestNextPublishTimePoint = nextPublishTimePoint;
-        }
-
-        if (timeUntilNextPublish.count() > 0) {
-          continue;
-        }
-
-        if (producerRegistration.hasQueueValue()) {
-          auto queueEntry = producerRegistration.popQueueValue();
-          publishMsg(queueEntry.data, queueEntry.topic,
-                     queueEntry.partitionNumber);
-        }
-
-        producerRegistration.resetLastPublishTimepoint();
-      }
-
-      auto minimalTimeUntilNextPublish =
-          closestNextPublishTimePoint - std::chrono::steady_clock::now();
-
-      if (minimalTimeUntilNextPublish.count() > 0) {
-        NDN_LOG_INFO("Sleeping for " << minimalTimeUntilNextPublish.count()
-                                     << " nanoseconds...");
-        std::this_thread::sleep_for(minimalTimeUntilNextPublish);
-      }
-    }
-
-    m_face.shutdown();
-
-    svsThread.join();
-  }
-
-  void shutdown() { m_running = false; }
+  void shutdown();
 
   friend IceflowConsumer;
   friend IceflowProducer;
@@ -151,79 +77,25 @@ public:
 private:
   uint32_t subscribeToTopicPartition(
       const std::string &topic, uint32_t partitionNumber,
-      std::function<void(std::vector<uint8_t>)> &pushDataCallback) {
+      std::function<void(std::vector<uint8_t>)> &pushDataCallback);
 
-    auto subscribedTopic = ndn::Name(topic).appendNumber(partitionNumber);
-
-    auto subscriptionHandle = m_svsPubSub->subscribe(
-        subscribedTopic, std::bind(&IceFlow::subscribeCallBack, this,
-                                   pushDataCallback, std::placeholders::_1));
-
-    std::cout << "Subscribed to " << topic << std::endl;
-
-    return subscriptionHandle;
-  }
-
-  void unsubscribe(uint32_t subscriptionHandle) {
-    m_svsPubSub->unsubscribe(subscriptionHandle);
-  }
+  void unsubscribe(uint32_t subscriptionHandle);
 
   void subscribeCallBack(
       const std::function<void(std::vector<uint8_t>)> &pushDataCallback,
-      const ndn::svs::SVSPubSub::SubscriptionData &subData) {
-    NDN_LOG_DEBUG("Producer Prefix: " << subData.producerPrefix << " ["
-                                      << subData.seqNo << "] : " << subData.name
-                                      << " : ");
-
-    std::vector<uint8_t> data(subData.data.begin(), subData.data.end());
-
-    pushDataCallback(data);
-  }
+      const ndn::svs::SVSPubSub::SubscriptionData &subData);
 
   void
-  onMissingData(const std::vector<ndn::svs::MissingDataInfo> &missing_data) {
-    // TODO: Implement if needed
-  }
+  onMissingData(const std::vector<ndn::svs::MissingDataInfo> &missing_data);
 
-  ndn::Name prepareDataName(const std::string &topic,
-                            uint32_t partitionNumber) {
-    return ndn::Name(topic).appendNumber(partitionNumber);
-  }
+  ndn::Name prepareDataName(const std::string &topic, uint32_t partitionNumber);
 
   void publishMsg(std::vector<uint8_t> payload, const std::string &topic,
-                  uint32_t partitionNumber) {
-    auto dataID = prepareDataName(topic, partitionNumber);
-    auto sequenceNo = m_svsPubSub->publish(
-        dataID, payload, ndn::Name(m_nodePrefix), ndn::time::seconds(4));
-    NDN_LOG_INFO("Publish: " << dataID << "/" << sequenceNo);
-  }
+                  uint32_t partitionNumber);
 
-  uint32_t registerProducer(ProducerRegistrationInfo producerRegistration) {
-    std::lock_guard lock(m_producerRegistrationMutex);
+  uint32_t registerProducer(ProducerRegistrationInfo producerRegistration);
 
-    uint32_t producerId = m_nextProducerId++;
-    m_producerRegistrations.insert({producerId, producerRegistration});
-
-    if (!m_producersAvailable) {
-      m_producersAvailable = true;
-      NDN_LOG_INFO(
-          "Producer has been registered, resuming producer procedure.");
-      m_producerRegistrationConditionVariable.notify_one();
-    }
-
-    return producerId;
-  }
-
-  void deregisterProducer(u_int64_t producerId) {
-    std::lock_guard lock(m_producerRegistrationMutex);
-
-    m_producerRegistrations.erase(producerId);
-
-    if (m_producerRegistrations.empty()) {
-      m_producersAvailable = false;
-      m_producerRegistrationConditionVariable.notify_one();
-    }
-  }
+  void deregisterProducer(u_int64_t producerId);
 
 private:
   ndn::KeyChain m_keyChain;
